Final fix: Simplified data types for serialization stability

+++ app/models/finance.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Numeric, ForeignKey
+from sqlalchemy import Column, Integer, String, Float
 from app.database import Base
 
 class ProjectFinancial(Base):
@@ -7,7 +7,7 @@ class ProjectFinancial(Base):
     id = Column(Integer, primary_key=True, index=True)
     project_id = Column(String)
     project_category = Column(String)
-    monthly_billing_amount = Column(Numeric)
+    monthly_billing_amount = Column(Float)
     billing_owner = Column(String)
     billing_start_date = Column(String)
     billing_end_date = Column(String)
@@ -18,7 +18,7 @@ class CostItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     cost_name = Column(String)
     cost_category = Column(String)
-    monthly_amount = Column(Numeric)
+    monthly_amount = Column(Float)
     owner_name = Column(String)
     start_date = Column(String)
     end_date = Column(String)
@@ -28,7 +28,7 @@ class Fund(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     investor_name = Column(String)
-    funding_amount = Column(Numeric)
+    funding_amount = Column(Float)
     funding_date = Column(String)
     funding_type = Column(String)
     responsible_owner = Column(String)
